Move sagas functions to separate files according to adr 0009
