+++ assets/js/state/databases.js
@@ -78,6 +78,14 @@ export const databasesListSlice = createSlice({
   },
 });
 
+export const DATABASE_REGISTERED = 'DATABASE_REGISTERED';
+export const DATABASE_HEALTH_CHANGED = 'DATABASE_HEALTH_CHANGED';
+export const DATABASE_INSTANCE_REGISTERED = 'DATABASE_INSTANCE_REGISTERED';
+export const DATABASE_INSTANCE_HEALTH_CHANGED =
+  'DATABASE_INSTANCE_HEALTH_CHANGED';
+export const DATABASE_INSTANCE_SYSTEM_REPLICATION_CHANGED =
+  'DATABASE_INSTANCE_SYSTEM_REPLICATION_CHANGED';
+
 export const {
   startDatabasesLoading,
   stopDatabasesLoading,
